feat(scheduling): Enhance booking flow with revamped calendar, time slot UI, and responsive layout
This commit delivers a full revamp of the scheduling experience across the booking flow.

Key changes:

Calendar:
- Feature: Date selection now highlights chosen dates with a green (success.main) background for instant feedback.
- Refactor: Simplified the multi-step scheduling flow into a single, reusable Calendar component.
- UI: Integrated a new MUI theme with Poppins font, purple/green palette, and glassmorphism Paper styling.
- Styling: Added gradient background, polished hover states, and improved button clarity for today, selected, and regular dates.

TimeSlot:
- Feature: Normalized slot handling for ISO strings and objects with availability states.
- UX: Responsive layout — full-width buttons on mobile, 2-column grid on larger screens.
- UI: Larger tap areas, adaptive font sizes, improved scrollable container with responsive height.
- Refactor: Clearer disabled state and consistent contained/outlined variants.

Booking Page:
- Refactor: Unified styling with consistent glass cards and gradient background.
- UI: Typography scales responsively, centered layout on mobile.
- UX: Improved step navigation with a clear 'Next Step' button at the bottom.
