+++ .env.development
@@ -1,1 +1,5 @@
 VITE_STRIPE_PUBLIC_KEY="pk_test_51OsCJ5P1A39VkufThp1PVDexesvf2XAY8faTyK0uucC1qRl9NW9QkpBdwXQDyjCAjzL166zjMWNn5Zr25ZkaQJVi00vurq61mj"
+VITE_API_BASE_URL=https://prevail-services-e973123f8b1e.herokuapp.com/api
+
+VITE_API_MEETING_SLOTS=/meeting/available-slots
+VITE_API_BOOK_MEETING=/meeting/book-meeting
